+++ src/com/udistrital/edu/model/ResultadoOrdenamiento.java
@@ -2,10 +2,10 @@ package com.udistrital.edu.model;
 
 public class ResultadoOrdenamiento {
     public double tiempo;
-    public int comparaciones;
-    public int intercambios;
+    public long comparaciones;
+    public long intercambios;
 
-    public ResultadoOrdenamiento(double tiempo, int comparaciones, int intercambios) {
+    public ResultadoOrdenamiento(double tiempo, long comparaciones, long intercambios) {
         this.tiempo = tiempo;
         this.comparaciones = comparaciones;
         this.intercambios = intercambios;
